Add more TUI tests and refactor.
