+++ datashuttle/tui/app.py
@@ -7,8 +7,8 @@ if TYPE_CHECKING:
 
 from pathlib import Path
 
+import showinfm
 import yaml
-from showinfm import show_in_file_manager
 from textual.app import App, ComposeResult
 from textual.containers import Container
 from textual.widgets import (
@@ -101,7 +101,7 @@ class TuiApp(App):
             return
 
         try:
-            show_in_file_manager(path_.as_posix())
+            showinfm.show_in_file_manager(path_.as_posix())
         except BaseException:
             if path_.is_file():
                 # I don't see why this is not working as according to docs it
